+++ lib/canvas/useOSMLSerializer.ts
@@ -1,13 +1,13 @@
 import { useCallback, useRef, useState } from "react";
 import type { ParsedElement } from "@/lib/canvas/types";
 import { useConfig } from "@/lib/config/ConfigProvider";
-import { OSMLSerializer } from "./osmlSerializer";
+import { OSMLStreamParser } from "./osmlStreamParser";
 
 const MAX_NODES_TO_SYNC = 3;
 
 export function useOSMLSerializer() {
 	const { connectorConfig } = useConfig();
-	const serializerRef = useRef(new OSMLSerializer());
+	const serializerRef = useRef(new OSMLStreamParser());
 	const [nodes, setNodes] = useState<ParsedElement[]>([]);
 
 	const appendChunk = useCallback(
